# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Chissl Server API
  version: 1.0.0
  description: Public and private APIs for the Chissl server. All endpoints are protected unless explicitly marked public.
servers:
  - url: /
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or API Token
  schemas:
...
    Listener:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        username: { type: string }
        port: { type: integer }
        mode: { type: string }
        use_tls: { type: boolean }
        status: { type: string }
        created_at: { type: string, format: date-time, nullable: true }
        updated_at: { type: string, format: date-time, nullable: true }
      additionalProperties: true
    ListenerList:
      type: array
      items: { $ref: '#/components/schemas/Listener' }
...
    ListenerCreateRequest:
      type: object
      required: [name, port, mode]
      properties:
        name: { type: string }
        port: { type: integer }
        mode: { type: string, enum: [sink, proxy, ai-mock] }
        target_url: { type: string }
        response: { type: string }
        use_tls: { type: boolean }
    ListenerUpdateRequest:
      type: object
      properties:
        name: { type: string }
        target_url: { type: string }
        response: { type: string }
        status: { type: string }
        use_tls: { type: boolean }
...
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /api/listeners:
    get:
      summary: List listeners
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListenerList'
    post:
      summary: Create listener
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListenerCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listener'
  /api/listener/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      summary: Get listener
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listener'
    put:
      summary: Update listener
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListenerUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listener'
    delete:
      summary: Delete listener
      responses: { '204': { description: No Content } }

  # Sessions & system (admin/user as indicated in code)
...